backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
google-generativeai
google-cloud-storage
uuid
python-multipart
langchain
langchain-community
langchain-google-genai
chromadb
beautifulsoup4
lxml
google-cloud-speech
websockets
httpx
sounddevice
numpy
